fix(frontend): clear React Doctor warnings on the PR's calendar files
- Header: hoist the pure segmentClass helper out of ZoomToggle to module
  scope as getZoomSegmentClass (prefer-module-scope-pure-function); rendered
  className is byte-identical.
- PhoneDayStrip: move buildPhoneDayStrip and its helpers/types into a sibling
  phoneDayStripModel.ts so the component file only exports components
  (only-export-components, Fast Refresh boundary).
Behaviour unchanged; react-doctor now reports zero warnings on this PR's files.
